+++ src/main/scala/esmeta/dump/visualizer/DumpNodeIdToProgId.scala
@@ -1,4 +1,4 @@
-package esmeta.dump
+package esmeta.dump.visualizer
 
 import esmeta.*
 import esmeta.cfg.*
